requirements.txt
requests
beautifulsoup4
pandas
google-cloud
google-cloud-pubsub
google-cloud-storage
